Crate gdk4_macos

Rust GDK 4 Macos bindings

The project website is here.

Rust bindings of GDK 4’s Macos backend, part of gtk4-rs.

GDK is an intermediate layer that isolates GTK from the details of the windowing system. GDK Macos contains functions specific to the Macos backend.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.83.

Documentation

The Rust API Stable/Development

The C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gdk-macos = { git = "https://github.com/gtk-rs/gtk4-rs.git", package = "gdk4-macos" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
gdk-macos = {version = "0.1", package = "gdk4-macos"}
gdk-wayland = { git = "https://github.com/gtk-rs/gtk4-rs.git", package = "gdk4-wayland" }

Features

Feature	Description

v4_8	Enable the new APIs part of GTK 4.8

cocoa	Integration with the cocoa crate

See Also

glib

gio

gsk4

gdk4

gtk4

License

The Rust bindings of gdk4-macos are available under the MIT License, please refer to it.

Re-exports

pub use gdk;

pub use gdk4_macos_sys as ffi;

pub use gio;

pub use glib;

Modules

prelude

Traits intended for blanket imports.

Structs

MacosDevice

Implements

MacosDisplay

Implements

MacosGLContext

Implements

MacosKeymap

GLib type: GObject with reference counted clone semantics.

MacosMonitor

Implements

MacosSeat

Implements

MacosSurface

This is an Abstract Base Class, you cannot instantiate it.

Type Aliases

idNon-cocoa